Make error handler configurable
