+++ snowhouse/assertmacro.h
@@ -9,7 +9,11 @@
 
 #include "assert.h"
 
+#ifndef SNOWHOUSE_NO_MACROS
+
 #define AssertThat(p1,p2)\
   Assert::That((p1), (p2), __FILE__, __LINE__);\
+
+#endif // SNOWHOUSE_NO_MACROS
 
 #endif	// IGLOO_ASSERTMACRO_H
