+++ snowhouse/fluent/operators/constraintoperator.h
@@ -7,22 +7,10 @@
 #ifndef IGLOO_CONTRAINTOPERATOR_H
 #define IGLOO_CONTRAINTOPERATOR_H
 
+#include "invalidexpressionexception.h"
+
 namespace snowhouse {
   
-  struct InvalidExpressionException
-  {
-    InvalidExpressionException(const std::string& message) : m_message(message)
-    {
-    }
-
-    const std::string& Message() const
-    {
-      return m_message;
-    }
-
-    std::string m_message;
-  };
-
   struct ConstraintOperator
   {
     virtual ~ConstraintOperator() {}
